move cross-val plot to its own module
